requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
httpx==0.28.1
pandas==2.3.2
numpy==2.3.2
scikit-learn==1.7.1
yfinance==0.2.65
plotly==6.3.0
streamlit==1.49.0
PyJWT==2.10.1
email-validator==2.3.0
langchain>=0.3,<0.4
langgraph>=0.2,<0.4
langchain-groq>=0.2,<0.4
langchain-chroma>=0.2,<0.3
chromadb>=0.5,<1.1
python-dotenv==1.0.1
slowapi>=0.1.9
pytest>=8.0
